+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix SpireField to not use duplicate objects
 * Fix NPE in isModLoaded
 * Force defining paramtypes on overloaded methods
+* Stricter error when method to patch isn't found
 
 #### v2.9.1 ####
 * Patch to always enable Custom mode